Refactoring the Templater parser class for more efficiency
